Add option to validate
when validate function is valled, it tries to convert given object
into configured type, and check constraints. However, it does not
convert object into configured type. This convertion should be
done in some place, so I think it's better to do it in joi.

To achive this, I added a option argument to validate. If convert
is true, validate converts given object to given configuration.

I believe this is useful when joi is used with hapi,
especially with query parsing.