Enable to pass "now" in date.min and date.max
It is not currently possible to validate a date relatively to the
current time. For example, the following schema will always validate
relatively to the time of the `schema` object allocation:

```javascript
var schema = {
  date: Joi.date().min(new Date())
};
```

This patch addresses this issue by adding a new `'now'` placeholder.
It dynamically computes the current time on-the-fly at each comparison.
Allowing to always compare relatively to the current time.

```javascript
var schema = {
  // Ensure the date field cannot represent a past date
  date: Joi.date().min('now')
};
```