(docs) Improve README with clarification on callbacks (#636)
Without cluttering up the README.md and keeping it as minimal as possible, added clarification that the callback is executed synchronously. This matches what is included in API.md and makes it clearer to new users without having to dig through the API docs or archived issues (#636) that there is no performance benefit of passing in a callback to validate, as typical node users would expect when such a callback option is provided.