Fix TypeScript types for strict object schema
A strict nested object was previously typed as `ObjectSchema<StrictSchemaMap<T>>`, which meant that Joi expected a Joi schema of a Joi schema of an object (i.e., a schema definition for an object of Joi schema fields) - which doesn't make much sense. Instead, it should require a Joi schema of an object _or_ an object of Joi schema fields.